use double's instead of int's to store ROI position rectangle

+++ MMSlideImage/src/org/bdgp/MMSlide/DB/ROI.java
@@ -6,10 +6,10 @@ import com.j256.ormlite.table.DatabaseTable;
 @DatabaseTable
 public class ROI {
     @DatabaseField(generatedId=true) private int id;
-    @DatabaseField private long x1;
-    @DatabaseField private long y1;
-    @DatabaseField private long x2;
-    @DatabaseField private long y2;
+    @DatabaseField private double x1;
+    @DatabaseField private double y1;
+    @DatabaseField private double x2;
+    @DatabaseField private double y2;
 
     public ROI() {}
     public ROI(int id, int imageId, long x1, long y1, long x2, long y2) {
@@ -20,8 +20,8 @@ public class ROI {
         this.y2 = y2;
     }
     public int getId() {return this.id;}
-    public long getX1() {return this.x1;}
-    public long getY1() {return this.y1;}
-    public long getX2() {return this.x2;}
-    public long getY2() {return this.y2;}
+    public double getX1() {return this.x1;}
+    public double getY1() {return this.y1;}
+    public double getX2() {return this.x2;}
+    public double getY2() {return this.y2;}
 }
